# packages/api-contracts/openapi.yaml
...
openapi: 3.0.0
paths:
  /api/v1/auth/register:
    post:
      description: Public. Returns a session.
      operationId: AuthController_register_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:
                  type: string
                  minLength: 1
                  maxLength: 120
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                  allOf:
                    - pattern: '[A-Z]'
                    - pattern: '[a-z]'
                    - pattern: '[0-9]'
              required:
                - fullName
                - email
                - password
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SessionResponseDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '422':
          description: BUSINESS_RULE_ERROR — the request was well-formed but violates a rule.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Create an account
      tags:
        - auth
  /api/v1/auth/login:
    post:
      description: Public.
      operationId: AuthController_login_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 1
              required:
                - email
                - password
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SessionResponseDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Sign in with email and password
      tags:
        - auth
  /api/v1/auth/google:
    post:
      description: Public.
      operationId: AuthController_google_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                idToken:
                  type: string
                  minLength: 1
                accessToken:
                  type: string
              required:
                - idToken
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SessionResponseDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Sign in with Google
      tags:
        - auth
  /api/v1/auth/refresh:
    post:
      description: Public — the refresh token itself is the credential.
      operationId: AuthController_refresh_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  minLength: 1
              required:
                - refreshToken
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/SessionResponseDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Exchange a refresh token for a new session
      tags:
        - auth
  /api/v1/auth/logout:
    post:
      description: Requires a bearer token.
      operationId: AuthController_logout_v1
      parameters:
        - name: authorization
          required: true
          in: header
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OkResultDto'
                required:
                  - success
                  - data
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Revoke the current session
      tags:
        - auth
  /api/v1/auth/forgot-password:
    post:
      description: Public. Always reports success — the response does not reveal whether the address is registered.
      operationId: AuthController_forgotPassword_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
              required:
                - email
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OkResultDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Send a password-reset email
      tags:
        - auth
...
info:
  title: Omnisift API
  description: Product Intelligence Platform — REST API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_ERROR
        - AUTHORIZATION_ERROR
        - NOT_FOUND
        - PROVIDER_ERROR
        - DATABASE_ERROR
        - BUSINESS_RULE_ERROR
        - INTERNAL_ERROR
      description: Stable machine-readable code. Safe to branch on.
    ApiErrorDetailDto:
      type: object
      properties:
        code:
          description: Stable machine-readable code. Safe to branch on.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe, user-facing message. Never contains a stack trace or internal detail.
      required:
        - code
        - message
    ApiErrorResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: false
          description: Always false on an error response.
        error:
          $ref: '#/components/schemas/ApiErrorDetailDto'
        requestId:
          type: string
          nullable: true
          description: Correlation id echoed from the x-request-id request header, or null.
      required:
        - success
        - error
        - requestId
    SessionResponseDto:
      type: object
      properties:
        userId:
          type: string
        email:
          type: string
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresAt:
          type: number
      required:
        - userId
        - email
        - accessToken
        - refreshToken
        - expiresAt
    OkResultDto:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required:
        - ok
...